make the check recursive.
