+++ aea/cli/add.py
@@ -31,9 +31,9 @@ from jsonschema import ValidationError
 
 from aea import AEA_DIR
 from aea.cli.common import Context, pass_ctx, logger, _try_to_load_agent_config
+from aea.cli.registry.utils import fetch_package, split_public_id
 from aea.configurations.base import DEFAULT_AEA_CONFIG_FILE, DEFAULT_CONNECTION_CONFIG_FILE, DEFAULT_SKILL_CONFIG_FILE, \
     DEFAULT_PROTOCOL_CONFIG_FILE
-from aea.cli.registry.utils import fetch_package, split_public_id
 
 
 @click.group()
